Created model and controllers for friends
