+++ models/friends.models.js
@@ -1,0 +1,24 @@
+const friends = [
+    {
+        id: 0,
+        name: 'Arijit Sahai'
+    },
+    {
+        id: 1,
+        name: 'Ujjwal Dhir'
+    },
+    {
+        id: 2,
+        name: 'Naveen Kumar'
+    },
+    {
+        id: 3,
+        name: 'Divyanand'
+    },
+    {
+        id: 4,
+        name: 'Pranjal Agarwal'
+    }
+]
+
+module.exports = friends;
